release: Chromium 25.0.1364 (25.0.1364.99)
Compare: https://github.com/chromium/chromium/compare/b852d81b4e79d80537e656bf4a6a9273b7fb13e0...9ca669a83526c914aca5f0e9fd75a10004f09043

+++ chrome/browser/tab_contents/spelling_menu_observer.h
@@ -9,9 +9,9 @@
 
 #include "base/compiler_specific.h"
 #include "base/memory/scoped_ptr.h"
+#include "base/prefs/public/pref_member.h"
 #include "base/string16.h"
 #include "base/timer.h"
-#include "chrome/browser/api/prefs/pref_member.h"
 #include "chrome/browser/tab_contents/render_view_context_menu_observer.h"
 
 class RenderViewContextMenuProxy;
@@ -75,7 +75,7 @@ class SpellingMenuObserver : public RenderViewContextMenuObserver {
   // service. The current animation just adds periods at the end of this string:
   //   'Loading' -> 'Loading.' -> 'Loading..' -> 'Loading...' (-> 'Loading')
   string16 loading_message_;
-  int loading_frame_;
+  size_t loading_frame_;
 
   // A flag represending whether a JSON-RPC call to the Spelling service
   // finished successfully and its response had a suggestion not included in the
